+++ components/Chat/components/ImageGallery/ImageGallery.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { MemoizedReactMarkdown } from '@/components/Markdown/MemoizedReactMarkdown';
 
 import MemoizedCarousel from '../Carousel/MemoizedCarousel';
-import MjImageComponent from '../MjImageComponent';
+import ImgComponent from './ImgComponent';
 
 import rehypeMathjax from 'rehype-mathjax';
 import rehypeRaw from 'rehype-raw';
@@ -36,44 +36,3 @@ export default function ImageGallery({ aiImageList }: ImageGalleryProps) {
     </MemoizedCarousel>
   );
 }
-
-const ImgComponent = ({
-  src,
-  node,
-}: React.DetailedHTMLProps<
-  React.ImgHTMLAttributes<HTMLImageElement> & { node?: any },
-  HTMLImageElement
->) => {
-  const isValidUrl = (url: string) => {
-    try {
-      new URL(url);
-      return true;
-    } catch (e) {
-      return false;
-    }
-  };
-
-  if (!src) return <></>;
-  if (!isValidUrl(src)) return <b>{`{InValid IMAGE URL}`}</b>;
-
-  const aiImageButtons =
-    node?.properties?.dataAiImageButtons &&
-    (node?.properties?.dataAiImageButtons).split(',');
-  const aiImagePrompt =
-    node?.properties?.dataAiImagePrompt &&
-    (node?.properties?.dataAiImagePrompt).split(',');
-  const aiImageButtonMessageId = node?.properties?.dataAiImageButtonMessageId;
-  if (aiImageButtons) {
-    return (
-      // eslint-disable-next-line @next/next/no-img-element
-      <MjImageComponent
-        src={src}
-        buttons={aiImageButtons}
-        buttonMessageId={aiImageButtonMessageId}
-        prompt={aiImagePrompt}
-      />
-    );
-  } else {
-    return <></>;
-  }
-};
